be extra sure of float32

+++ tiktorch/handler/inference.py
@@ -222,7 +222,7 @@ class InferenceProcess(IInference):
             end = next(end_generator)
             try:
                 with torch.no_grad():
-                    pred = model(torch.stack(data[start:end]).to(device=device)).cpu()
+                    pred = model(torch.stack(data[start:end]).to(dtype=torch.float, device=device)).cpu()
             except Exception as e:
                 if batch_size > last_batch_size:
                     self.logger.info(
